Delete TipView's channel count getters in favor of those in PipetteView.
There is one behavioral change here, from the perspective of a protocol engine command implementation.

Formerly, the initial value of get_active_channels(), after the pipette was loaded and the first PipetteConfigUpdate was applied to state, would be based on the pipette's *total* channels. This looked like a bug. It is now based on the active channels declared in PipetteConfigUpdate.config.nozzle_map.

Because loadPipette commands always start pipettes off in their full nozzle configuration, those two values have always agreed in practice. Therefore, this internal fix does not change anything from the perspective of a protocol engine command issuer. It's just more theoretically correct.
