+++ api/src/opentrons/protocol_engine/commands/get_next_tip.py
@@ -73,7 +73,7 @@ class GetNextTipImplementation(
         pipette_id = params.pipetteId
         starting_tip_name = params.startingTipWell
 
-        num_tips = self._state_view.tips.get_pipette_active_channels(pipette_id)
+        num_tips = self._state_view.pipettes.get_active_channels(pipette_id)
         nozzle_map = self._state_view.pipettes.get_nozzle_configuration(pipette_id)
 
         if (
