+++ api/tests/opentrons/protocol_engine/commands/test_get_next_tip.py
@@ -25,7 +25,7 @@ async def test_get_next_tip_implementation(
     )
     mock_nozzle_map = decoy.mock(cls=NozzleMap)
 
-    decoy.when(state_view.tips.get_pipette_active_channels("abc")).then_return(42)
+    decoy.when(state_view.pipettes.get_active_channels("abc")).then_return(42)
     decoy.when(state_view.pipettes.get_nozzle_configuration("abc")).then_return(
         mock_nozzle_map
     )
@@ -60,7 +60,7 @@ async def test_get_next_tip_implementation_multiple_tip_racks(
     )
     mock_nozzle_map = decoy.mock(cls=NozzleMap)
 
-    decoy.when(state_view.tips.get_pipette_active_channels("abc")).then_return(42)
+    decoy.when(state_view.pipettes.get_active_channels("abc")).then_return(42)
     decoy.when(state_view.pipettes.get_nozzle_configuration("abc")).then_return(
         mock_nozzle_map
     )
@@ -95,7 +95,7 @@ async def test_get_next_tip_implementation_no_tips(
     )
     mock_nozzle_map = decoy.mock(cls=NozzleMap)
 
-    decoy.when(state_view.tips.get_pipette_active_channels("abc")).then_return(42)
+    decoy.when(state_view.pipettes.get_active_channels("abc")).then_return(42)
     decoy.when(state_view.pipettes.get_nozzle_configuration("abc")).then_return(
         mock_nozzle_map
     )
@@ -124,7 +124,7 @@ async def test_get_next_tip_implementation_partial_with_starting_tip(
     )
     mock_nozzle_map = decoy.mock(cls=NozzleMap)
 
-    decoy.when(state_view.tips.get_pipette_active_channels("abc")).then_return(42)
+    decoy.when(state_view.pipettes.get_active_channels("abc")).then_return(42)
     decoy.when(state_view.pipettes.get_nozzle_configuration("abc")).then_return(
         mock_nozzle_map
     )
